Attempt at removing race condition on email row insertion
the deleted get_first_invoice wasn't an atomic database op, so it was
happening to hit a non-unique payment-hash in the following email row insert.

Can't find how to do a db transaction and doing a join query requires changes
in the db model.

The workaround is to return a list of invoices, trying to insert an email with
each minimizing the chance of finding another thread with the same payment hash